+++ fidimag/common/integrators.py
@@ -3,7 +3,7 @@ Implement the time integrators used in the simulation of magnetisation dynamics.
 
 """
 from scipy.integrate import ode
-from fidimag.extensions.cvode import CvodeSolver, CvodeSolver_OpenMP
+from fidimag.extensions.cvode import CvodeSolver, CvodeSolver_OpenMP, ErkSolver
 import warnings
 
 EPSILON = 1e-16